ADD first iteration of a mock log in page.
quite happy with knowledge obtain, but not quite happy with the clean and lean of the codes. will need to try to use more variables for colours, as well as for containers. to make code more clean and consistent

will not re-visit the task, maybe will make a new one with original ideas, as much of this layout and editing stem from few website that has login pages that i like.